api/requirements.txt
fastapi==0.113.0
uvicorn==0.30.6
pymongo==4.9.1
elasticsearch==8.15.0
python-dotenv==1.0.1
pydantic==2.9.2
